tweaks to get additional pages working
