+++ app/views/branded/plans/index.html.erb
@@ -3,7 +3,9 @@
   <div class="col-md-12">
     <h1>
       <%= _('My Dashboard') %>
-      <%= link_to _('Dashboard Prototype'), dashboard_path, class: 'btn btn-primary pull-right' %>
+      <% if current_user.can_org_admin? %>
+        <%= link_to _('Dashboard Prototype'), dashboard_path, class: 'btn btn-primary pull-right' %>
+      <% end %>
     </h1>
 
     <!-- if the user has projects -->
